Challenge completed
Added a main menu. Player is able to input it's name. If there is no name input - game cannot continue. Player's name is saved between scenes and the best score (along with the name of the player who set it) saves between sessions and is displayed at the bottom of the screen (when playing and if available)